Upgraded Tracking Algorithm

+++ STCore/utils/backgroundEstimator.py
@@ -8,7 +8,7 @@ def GetBackground(image):
     #startTime = time()
     #bkg_estimator = ModeEstimatorBackground()
     # bkg = Background2D(image, (20, 30), filter_size=(4, 4),sigma_clip=sigma_clip, bkg_estimator=bkg_estimator)
-    background = mean(mode(image)[0])
+    background = int(mean(mode(image)[0]))
     #print "Tiempo: ", time() - startTime
     #plt.imshow(bkg.background, origin='lower', cmap='Greys_r')
     #background=bkg.bkg_estimator.calc_background(image)
